Move chain-related code to dedicated files
